requirements.txt
numpy
pandas
scikit-learn
matplotlib
pillow
opencv-python
streamlit
torch
torchvision
tensorflow
transformers
sentencepiece
scikit-image
joblib
